Fix bug caused by colormap setting
Previously,  when a colormap was set, it was set for the current figure, which may affect other axes in that figure. Targets of this operation are now correctly set in three functions run in the main dialog. No change to other corresponding functions because they run in a separate window.