Commandbar icon switching on/off vectorization
"norm" is a little bit long and cannot be clearly written using 16X16 pixels.
Thus, only "Vec." was displayed on the button, but it can switch the status between black and gray, indicating on/off states.
The background color can now correctly changes also, because the transparency has been defined.